+++ lib/models/field_types/convienence_classes/checkboxselect.dart
@@ -1,5 +1,6 @@
 import 'package:championforms/models/field_builder_context.dart';
 import 'package:championforms/core/field_json.dart';
+import 'package:championforms/models/field_colors.dart';
 import 'package:championforms/models/field_condition.dart';
 import 'package:championforms/models/field_types/optionselect.dart';
 import 'package:championforms/models/multiselect_option.dart';
@@ -35,6 +36,8 @@ class CheckboxSelect extends OptionSelect {
     super.onChange,
     super.fieldLayout,
     super.fieldBackground,
+    super.colors,
+    super.animateValidationErrors,
     Widget Function(FieldBuilderContext)? fieldBuilder,
   }) : super(fieldBuilder: fieldBuilder ?? checkboxFieldBuilder);
 
@@ -80,6 +83,8 @@ class CheckboxSelect extends OptionSelect {
     Widget? leading,
     Widget? trailing,
     FormTheme? theme,
+    FieldColors? colors,
+    bool? animateValidationErrors,
     String? title,
     String? description,
     bool? disabled,
@@ -116,6 +121,9 @@ class CheckboxSelect extends OptionSelect {
       leading: leading ?? this.leading,
       trailing: trailing ?? this.trailing,
       theme: theme ?? this.theme,
+      colors: colors ?? this.colors,
+      animateValidationErrors:
+          animateValidationErrors ?? this.animateValidationErrors,
       title: title ?? this.title,
       description: description ?? this.description,
       disabled: disabled ?? this.disabled,