+++ lib/models/field_types/formfielddefnull.dart
@@ -1,3 +1,4 @@
+import 'package:championforms/models/field_colors.dart';
 import 'package:championforms/models/field_types/formfieldclass.dart';
 import 'package:championforms/models/file_model.dart';
 import 'package:championforms/models/themes.dart';
@@ -28,6 +29,8 @@ class NullField extends Field {
     super.onChange,
     super.fieldLayout,
     super.fieldBackground,
+    super.colors,
+    super.animateValidationErrors,
     this.defaultValue,
   });
 
@@ -59,6 +62,8 @@ class NullField extends Field {
       FieldColorScheme currentColors,
       Widget renderedField,
     )? fieldBackground,
+    FieldColors? colors,
+    bool? animateValidationErrors,
     String? defaultValue,
   }) {
     return NullField(
@@ -76,6 +81,9 @@ class NullField extends Field {
       onChange: onChange ?? this.onChange,
       fieldLayout: fieldLayout ?? this.fieldLayout,
       fieldBackground: fieldBackground ?? this.fieldBackground,
+      colors: colors ?? this.colors,
+      animateValidationErrors:
+          animateValidationErrors ?? this.animateValidationErrors,
       defaultValue: defaultValue ?? this.defaultValue,
     );
   }