+++ lib/models/field_types/optionselect.dart
@@ -1,5 +1,6 @@
 import 'package:championforms/models/field_builder_context.dart';
 import 'package:championforms/core/field_json.dart';
+import 'package:championforms/models/field_colors.dart';
 import 'package:championforms/models/field_condition.dart';
 import 'package:championforms/models/field_types/formfieldclass.dart';
 import 'package:championforms/models/file_model.dart';
@@ -74,6 +75,8 @@ class OptionSelect extends Field {
     super.onChange,
     super.fieldLayout,
     super.fieldBackground,
+    super.colors,
+    super.animateValidationErrors,
     Widget Function(FieldBuilderContext)? fieldBuilder,
   }) : fieldBuilder = fieldBuilder ?? dropdownFieldBuilder;
 
@@ -118,6 +121,8 @@ class OptionSelect extends Field {
     Widget? leading,
     Widget? trailing,
     FormTheme? theme,
+    FieldColors? colors,
+    bool? animateValidationErrors,
     String? title,
     String? description,
     bool? disabled,
@@ -154,6 +159,9 @@ class OptionSelect extends Field {
       leading: leading ?? this.leading,
       trailing: trailing ?? this.trailing,
       theme: theme ?? this.theme,
+      colors: colors ?? this.colors,
+      animateValidationErrors:
+          animateValidationErrors ?? this.animateValidationErrors,
       title: title ?? this.title,
       description: description ?? this.description,
       disabled: disabled ?? this.disabled,